+++ NEWS.md
@@ -5,7 +5,7 @@
   * `check_package()` will warn if the class is missing.
 * `read_resource()` no longer returns a message regarding rights and credit (#121). If `package$id` is a URL (e.g. a DOI) it will be mentioned in `print()`.
 * `read_resource()` now supports column selection via the `col_select` argument from `readr::read_delim()`. This can vastly improve reading speed (#123).
-* `add_resource()` now supports adding additional resource properties via the `...` argument.
+* `add_resource()` now supports adding additional resource properties via the `...` argument. These are retained in `write_package()` (#195).
 * `create_package()` now accepts a `descriptor` argument so that a Data Package object can be created from an existing object (#184). It will always validate the created object with `create_package()`.
 * `check_package()` is now a public function, so it can be used by other packages (#185). This and the other `check_` functions now return the first argument silently (rather than `TRUE`), so they can be chained.
 * `readr::problems()` is included in NAMESPACE so users don't have to load readr to inspect parsing issues. The function is mentioned in the documentation of `read_resource()` (#129).